Añadido el back para la obtención del menú

+++ src/main/java/com/ebarapp/ebar/repository/BarRepository.java
@@ -1,11 +1,9 @@
 package com.ebarapp.ebar.repository;
 
-
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.ebarapp.ebar.model.Bar;
 
-public interface BarRepository extends JpaRepository<Bar, Long> {
+public interface BarRepository extends JpaRepository<Bar, Integer> {
 
 }
